cms/requirements.txt
apturl==0.5.2
asgiref==3.3.1
astroid==2.5.1
blinker==1.4
Brlapi==0.7.0
certifi==2019.11.28
chardet==3.0.4
Click==7.0
colorama==0.4.3
command-not-found==0.3
cryptography==2.8
cupshelpers==1.0
dbus-python==1.2.16
defer==1.0.6
distro==1.4.0
distro-info===0.23ubuntu1
Django==3.1.7
django-filter==2.4.0
django-tts==0.3.1
djangorestframework==3.12.2
dnspython==1.16.0
docutils==0.16
drf-extensions==0.7.0
drf-nested-routers==0.93.3
entrypoints==0.3
httplib2==0.14.0
idna==2.8
isort==5.7.0
keyring==18.0.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lazy-object-proxy==1.5.2
louis==3.12.0
macaroonbakery==1.3.1
Markdown==3.3.4
mccabe==0.6.1
netifaces==0.10.4
oauthlib==3.1.0
olefile==0.46
pexpect==4.6.0
Pillow==7.0.0
protobuf==3.6.1
pycairo==1.16.2
pycrypto==2.6.1
pycups==1.9.73
PyGObject==3.36.0
PyJWT==1.7.1
pylint==2.7.2
pymacaroons==0.13.0
PyNaCl==1.3.0
pyRFC3339==1.1
python-apt==2.0.0+ubuntu0.20.4.4
python-dateutil==2.7.3
python-debian===0.1.36ubuntu1
pyttsx3==2.90
pytz==2019.3
pyxdg==0.26
PyYAML==5.3.1
reportlab==3.5.34
requests==2.22.0
requests-unixsocket==0.2.0
rhvoice-wrapper==0.6.3
SecretStorage==2.3.1
simplejson==3.16.0
six==1.14.0
sqlparse==0.4.1
ssh-import-id==5.10
systemd-python==234
toml==0.10.2
ubuntu-advantage-tools==20.3
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.25.8
wadllib==1.3.3
wrapt==1.12.1
xkit==0.0.0
